refactored brigtness sensors to use floats

+++ NewMotorDriver/src/lineDetector/brightnessSensor/BrightnessSensor.hpp
@@ -9,7 +9,7 @@ public:
     BrightnessSensor(){};
     ~BrightnessSensor(){};
 
-    virtual uint8_t getValues(uint8_t *array, uint8_t maxLen) = 0;
+    virtual uint8_t getValues(float *array, uint8_t maxLen) = 0;
     virtual uint8_t numberOfSensors() = 0;
 };
